// src/shared/generated/EmbeddedKeyEffectRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type EmbeddedKeyEffectRecord = { phase: "rawKeyDown" | "keyUp", code: string, activeCodesBefore: Array<string>, activeCodes: Array<string>, autoRepeat: boolean, suppressShortcut: boolean, };