// src/shared/generated/EmbeddedKeyTransitionRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EmbeddedKeyEffectRecord } from "./EmbeddedKeyEffectRecord";

export type EmbeddedKeyTransitionRecord = { transitionId?: string, effects: Array<EmbeddedKeyEffectRecord>, hasHeldKeys: boolean, };